requirements.txt
python-evtx==0.7.4
python-registry==1.3.1
pywin32==306
